+++ spec/integrations/shoryuken_spec.rb
@@ -5,7 +5,8 @@ describe 'Bugsnag::Shoryuken', :order => :defined do
   before do
     unless defined?(::Shoryuken)
       @mocked_shoryuken = true
-      class ::Shoryuken
+      class Shoryuken
+        VERSION = '1.2.3'
       end
       module Kernel
         alias_method :old_require, :require
@@ -27,36 +28,40 @@ describe 'Bugsnag::Shoryuken', :order => :defined do
   end
 
   it "calls configure when initialised" do
-    config = double("config")
-
-    expect(config).to receive(:detected_app_type=).with("shoryuken")
-    expect(config).to receive(:default_delivery_method=).with(:synchronous)
-    expect(Bugsnag).to receive(:configure).and_yield(config)
     Bugsnag::Shoryuken.new
+
+    expect(Bugsnag.configuration.app_type).to eq("shoryuken")
+    expect(Bugsnag.configuration.delivery_method).to eq(:synchronous)
+    expect(Bugsnag.configuration.runtime_versions["shoryuken"]).to eq(Shoryuken::VERSION)
   end
 
   it "calls correct sequence when called" do
-    queue = 'queue'
-    body = 'body'
+    queue = 'a queue name'
+    body = 'the body of a queued message'
+    exception = RuntimeError.new('oops')
 
-    callbacks = double('callbacks')
-    expect(callbacks).to receive(:<<) do |func|
-      report = double('report')
-      expect(report).to receive(:add_tab).with(:shoryuken, {
-        queue: queue,
-        body: body
+    expect {
+      Bugsnag::Shoryuken.new.call('_', queue, '_', body) { raise exception }
+    }.to raise_error(exception)
+
+    expect(Bugsnag).to have_sent_notification { |payload, headers|
+      event = get_event_from_payload(payload)
+
+      expect(event['unhandled']).to be(true)
+      expect(event['severity']).to eq('error')
+      expect(event['app']['type']).to eq('shoryuken')
+      expect(event['device']['runtimeVersions']['shoryuken']).to eq(Shoryuken::VERSION)
+
+      expect(event['metaData']['shoryuken']).to eq({
+        'body' => body,
+        'queue' => queue,
       })
-      func.call(report)
-    end
-    config = double('config')
-    allow(config).to receive(:detected_app_type=).with("shoryuken")
-    allow(config).to receive(:default_delivery_method=).with(:synchronous)
-    allow(config).to receive(:clear_request_data)
-    expect(Bugsnag).to receive(:before_notify_callbacks).and_return(callbacks)
-    allow(Bugsnag).to receive(:configure).and_yield(config)
-    allow(Bugsnag).to receive(:configuration).and_return(config)
-    shoryuken = Bugsnag::Shoryuken.new
-    expect { |b| shoryuken.call('_', queue, '_', body, &b )}.to yield_control
+
+      expect(event['severityReason']).to eq({
+        'type' => Bugsnag::Report::UNHANDLED_EXCEPTION_MIDDLEWARE,
+        'attributes' => { 'framework' => 'Shoryuken' }
+      })
+    }
   end
 
   after do